Create a flexible TLS server abstraction
This project needs to be run in 2 different ways:

 * Directly in a customer environment (typical TLS requirements)
 * multi-tenant K8s in foundries backend (like our device-gateway and
   ostree-hub).

In the multi-tenant case, we have special logic to look at the incoming
SNI information in the request and use the correct certs for a factory.

This commit introduces the abstraction and simple single-tenant
implementation. Foundries can then extend this project with their own
handler to satisfy their requirements.

Signed-off-by: Andy Doan <andy@foundries.io>